// src/api/generated/models/getCompaniesCompanyIdRequestLinesParams.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */
import type { GetCompaniesCompanyIdRequestLinesStatus } from './getCompaniesCompanyIdRequestLinesStatus';
import type { GetCompaniesCompanyIdRequestLinesUndistributed } from './getCompaniesCompanyIdRequestLinesUndistributed';
import type { GetCompaniesCompanyIdRequestLinesWithoutQuotes } from './getCompaniesCompanyIdRequestLinesWithoutQuotes';
import type { GetCompaniesCompanyIdRequestLinesSortBy } from './getCompaniesCompanyIdRequestLinesSortBy';
import type { GetCompaniesCompanyIdRequestLinesSortOrder } from './getCompaniesCompanyIdRequestLinesSortOrder';

export type GetCompaniesCompanyIdRequestLinesParams = {
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
/**
 * @minimum 0
 * @maximum 9007199254740991
 */
offset?: number;
status?: GetCompaniesCompanyIdRequestLinesStatus;
/**
 * @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
 */
createdByUserId?: string;
/**
 * @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
 */
requestId?: string;
/**
 * @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
 */
productId?: string;
/**
 * @minLength 1
 */
q?: string;
undistributed?: GetCompaniesCompanyIdRequestLinesUndistributed;
withoutQuotes?: GetCompaniesCompanyIdRequestLinesWithoutQuotes;
sortBy?: GetCompaniesCompanyIdRequestLinesSortBy;
sortOrder?: GetCompaniesCompanyIdRequestLinesSortOrder;
};